requirements.txt
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
crypto==1.4.1
cryptography==36.0.1
et-xmlfile==1.1.0
h11==0.12.0
idna==2.10
iniconfig==1.1.1
Naked==0.1.31
numpy==1.22.0
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
panda==0.3.1
pandas==1.3.5
pdfminer.six==20211012
pdfplumber==0.6.0
Pillow==9.0.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.25.1
selenium==4.1.0
shellescape==3.8.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
style==1.1.0
toml==0.10.2
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
update==0.0.1
urllib3==1.26.5
Wand==0.6.7
wsproto==1.0.0
